requirements.txt
streamlit
langchain
openai
faiss-cpu
unstructured
python-dotenv
requests
python-docx
pypdf